Test w/ `mini_httpd`: Use `jau::io::uri_tk::protocol_supported("http:")` instead of `#ifdef USE_LIBCURL`; Fix test package for TestByteStream01 and fix Clock.cxx name

+++ test/test_iostream01.cpp
@@ -67,21 +67,21 @@ class TestIOStream01 {
                     ofs.write(reinterpret_cast<char*>(one_line.data()), one_line.size());
                 }
             }
-#ifdef USE_LIBCURL
-            int res = std::system("killall mini_httpd");
-            const std::string cwd = jau::fs::get_cwd();
-            const std::string cmd = "/usr/sbin/mini_httpd -p 8080 -l "+cwd+"/mini_httpd.log";
-            jau::PLAIN_PRINT(true, "%s", cmd.c_str());
-            res = std::system(cmd.c_str());
-            (void)res;
-#endif // USE_LIBCURL
+            if( jau::io::uri_tk::protocol_supported("http:") ) {
+                int res = std::system("killall mini_httpd");
+                const std::string cwd = jau::fs::get_cwd();
+                const std::string cmd = "/usr/sbin/mini_httpd -p 8080 -l "+cwd+"/mini_httpd.log";
+                jau::PLAIN_PRINT(true, "%s", cmd.c_str());
+                res = std::system(cmd.c_str());
+                (void)res;
+            }
         }
 
         ~TestIOStream01() {
-#ifdef USE_LIBCURL
-            int res = std::system("killall mini_httpd");
-            (void)res;
-#endif // USE_LIBCURL
+            if( jau::io::uri_tk::protocol_supported("http:") ) {
+                int res = std::system("killall mini_httpd");
+                (void)res;
+            }
         }
 
         void test00_protocols() {
